Move PersonForm tests from People.test.tsx to PersonForm.test.tsx
